fasttools-mobile-app/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
requests
pillow
pydub
sentry-sdk
python-dotenv
